% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a Number}
\usage{
square(x)
}
\arguments{
\item{x}{Number to square}
}
\value{
The given number to the second power
}
\description{
Square a Number
}
\examples{
square(3)
}
